requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.4
django-amplitude==0.7.0
django-static-jquery==2.1.4
gensim==3.8.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
kombu==5.1.0
nltk==3.6.2
numpy==1.19.3
pandas==1.2.5
prompt-toolkit==3.0.19
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.7.1
rfc3986==1.5.0
scipy==1.7.0
six==1.16.0
smart-open==5.1.0
sniffio==1.2.0
sqlparse==0.4.1
tqdm==4.61.1
ua-parser==0.10.0
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
